kyc-ai/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
pdfplumber>=0.10.2
docx2txt>=0.8